pythonProject1/pyproject.toml
[tool.poetry]
name = "finallywork"
version = "0.1.0"
description = ""
authors = ["npoblema"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pandas = "^2.2.2"
yfinance = "^0.2.40"
xlrd = "^2.0.1"
python-dotenv = "^1.0.1"
types-requests = "^2.32.0.20240622"


[tool.poetry.group.lint.dependencies]
mypy = "^1.9.0"
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
autopep8 = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
line_length = 120
profile = "black"
multi_line_output = 3

[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
)
)
'''
